+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.4.0
+
+`2023-03-17`
+
+### Features
+
+- New Serie Download Options
+  - rename: Override the serie title, for CLI: reusing -n, --name
+  - retry: Automatically redownload failed chapters, for CLI: -r, --retry
+  - chapters: Only downloading given list of chapter indexes, for CLI: -c, --chapters 1,2,4,7
+- Better logs when partially downloading series
+
 ## 1.3.0
 
 `2023-03-15`